% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup_mse_options.R
\docType{data}
\name{abc_regionfleet_allocation}
\alias{abc_regionfleet_allocation}
\title{ABC Allocation Rate to Fleets within Regions}
\format{
An object of class \code{matrix} (inherits from \code{array}) with 5 rows and 2 columns.
}
\usage{
abc_regionfleet_allocation
}
\description{
Values represent the proportion of the regional ABC that is allocated to each fleet.
Values come from the GOA Groudfish FMP and BSAI Groundfish FMP.
}
\keyword{datasets}
