% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/age_structure_stats.R
\name{abi}
\alias{abi}
\title{Compute Age-Based Indicator}
\usage{
abi(naa, ref, threshold = 0.9, start_age = 2)
}
\arguments{
\item{naa}{numbers-at-age vector}

\item{ref}{a reference naa vector}

\item{threshold}{proportion of population to consider "old"}

\item{start_age}{age at which to start calculating ABI}
}
\description{
Compute ABI (Griffiths et al. 2023) given a
population age structure and an age structure
to use as a reference.
}
