% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregate_observations.R
\name{aggregate_indices}
\alias{aggregate_indices}
\title{Aggregate Survey Index Observations}
\usage{
aggregate_indices(survey_obs, y)
}
\arguments{
\item{survey_obs}{list of survey observations that include an RPN and RPW elements of
dimension [nyears, 1, 1, nregions, nsurveys]}

\item{y}{the terminal year of survey observations to perform aggregation on}
}
\value{
list of RPN and RPW indices aggregated to a single spatial dimension
}
\description{
Given observations of survey indices generated by the OM (afscOM::simulate_observations)
across more than one region, aggregate such indices to a single region by summing
across regions.
}
