% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/harvest_control_rules.R
\name{as_scalar_threshold_f}
\alias{as_scalar_threshold_f}
\title{Threshold HCR with Fmax scaled by Age Structure}
\usage{
as_scalar_threshold_f(x, naa, ref_naa, as_func, ..., f_min, f_max, lrp, urp)
}
\arguments{
\item{x}{input to HCR (e.g. SSB or, SSB/B40)}

\item{naa}{an age structure (e.g. numbers at age or biomass at age)}

\item{ref_naa}{am age structure to compare `naa` to}

\item{as_func}{a function that compute an age structure statistic (must take naa as its first argument)}

\item{...}{additional arguments to pass to `as_func(naa, ...)`}

\item{f_min}{F to use when x < lrp}

\item{f_max}{F to use when x > urp}

\item{lrp}{value of x below which f_min applies}

\item{urp}{value of x above which f_max applies}
}
\description{
Compute an allowable fishing mortality rate from
a threshold HCR parameterized by f_min, f_max, lrp,
and urp, but scales f_max by an age structure summary
statistic relative to the value of that summmary
statistic for some reference age structure.
}
