% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/age_structure_stats.R
\name{average_age}
\alias{average_age}
\title{Compute average age of a population
#'
Calculate the average age of a population
given its age structure.}
\usage{
average_age(naa, ages)
}
\arguments{
\item{naa}{numbers-at-age vector}

\item{ages}{vector of ages}
}
\description{
Compute average age of a population
#'
Calculate the average age of a population
given its age structure.
}
