% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/performance_metrics.R
\name{average_annual_value}
\alias{average_annual_value}
\title{Compute Average Annual Value of Catch across projection period}
\usage{
average_annual_value(
  model_runs,
  extra_columns,
  interval_widths = c(0.5, 0.8),
  time_horizon = c(65, NA),
  extra_filter = NULL,
  relative = NULL,
  summarise_by = c("om", "hcr")
)
}
\arguments{
\item{model_runs}{list of completed MSE simulations runs}

\item{extra_columns}{data.frame specifying names for OM and HCR to attach
to each model_run (see `bind_mse_outputs` for more details)}

\item{interval_widths}{confidence intevrals to compute}

\item{extra_filter}{an additional set of filters to apply before computing 
medians and confidence intervals}

\item{relative}{a management procedure to compute metric relative to}

\item{summarise_by}{vector of columns to summarise metric by}

\item{summary_out}{whether to output data summarised by `ggdist` or full data}
}
\description{
Compute average annual value of landed catch (median and CIs) 
per year across all years and simulation seeds, for each combination 
of operating models and management procedures. Fish >9yo are considered
"high value" and are assigned a relative value of 1, fish 5-9yo are
considered "medium value" and are asisgned a relative value of 0.556, 
and fish <5yo are considered "low value" and are assigned a relative
valye of 0.337.
}
