% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attainment_utils.R
\docType{data}
\name{average_regflt_tac_utilization}
\alias{average_regflt_tac_utilization}
\title{Average Region-Fleet TAC Utilization}
\format{
An object of class \code{matrix} (inherits from \code{array}) with 5 rows and 2 columns.
}
\usage{
average_regflt_tac_utilization
}
\description{
Average Region-Fleet TAC utilization rates from 2000-2020 as derived
from historical regional ABCs, fleet allocations, and region-specific
landings by fleet. See `dev/historical_TAC_analyses.R`.
}
\keyword{datasets}
