% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recruitment_utils.R
\name{beverton_holt}
\alias{beverton_holt}
\title{Beverton-Holt stock recruitment relationship
#'
Defines a function factory for yieding recruitment from a Bevrton-Holt
stock recruitment relationship, parameterized using steepeness. Initial
call to this function sets up a new function that takes the current SSB
as input, and returns the predicted recruitment from the BH SRR defined
by h, R0, and S0.}
\usage{
beverton_holt(h, R0, S0, sigR, seed)
}
\arguments{
\item{h}{steepness (0.2 <= h <= 1.0)}

\item{R0}{unfished recruitment}

\item{S0}{unfished spawning biomass}
}
\description{
Beverton-Holt stock recruitment relationship
#'
Defines a function factory for yieding recruitment from a Bevrton-Holt
stock recruitment relationship, parameterized using steepeness. Initial
call to this function sets up a new function that takes the current SSB
as input, and returns the predicted recruitment from the BH SRR defined
by h, R0, and S0.
}
\examples{
\dontrun{
 bevholt <- beverton_holt(0.7, 25, 300)
 bevholt(ssb=185) # 23.43891
}

}
