% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recruitment_utils.R
\name{bevholt_regimes}
\alias{bevholt_regimes}
\title{Regime-like Recruitment with Beverton-Holt SRR}
\usage{
bevholt_regimes(
  h,
  sbpr,
  R0s,
  sigRs,
  nyears,
  regime_length,
  starting_regime,
  seed
)
}
\arguments{
\item{h}{steepness of SRR (0.2 <= h <= 1.0)}

\item{sbpr}{spawning biomass per recruit}

\item{R0s}{vector of unfished recruitment in each regime}

\item{sigRs}{vector of recruitment variability in each regime}

\item{nyears}{total number of years to simulate recruitment}

\item{regime_length}{vector inidcating how long each regime lasts (limited to 2 regimes)}

\item{starting_regime}{regime to start with (0 = first regime, 1 = second regime)}

\item{seed}{random seed for reproducability}
}
\description{
Generate future recruitment timeseries based on two distincr Beverton-Holt SRRs,
that difer based on R0 or sigR.
}
