% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_utils.R
\name{bind_mse_outputs}
\alias{bind_mse_outputs}
\title{Create Tibble of MSE Output Data from Multiple Models}
\usage{
bind_mse_outputs(model_runs, var, extra_columns)
}
\arguments{
\item{model_runs}{a list of MSE model run objects (created via `run_mse(...)`)}

\item{var}{the output variable from the model run objects}

\item{extra_columns}{a data.frame of extra column names and values to add to the tibble}
}
\description{
Create Tibble of MSE Output Data from Multiple Models
}
