% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/performance_metrics.R
\name{biomass_recovery_time}
\alias{biomass_recovery_time}
\title{Compute number of years requied for SSB to recover}
\usage{
biomass_recovery_time(
  model_runs,
  extra_columns,
  dem_params,
  hcr_filter,
  om_filter,
  interval_widths = c(0.5, 0.8),
  time_horizon = c(55, NA),
  extra_filter = NULL,
  relative = NULL,
  summarise_by = c("om", "hcr"),
  summary_out = TRUE
)
}
\arguments{
\item{model_runs}{list of completed MSE simulations runs}

\item{extra_columns}{data.frame specifying names for OM and HCR to attach
to each model_run (see `bind_mse_outputs` for more details)}

\item{hcr_filter}{vector of HCR names to calculate metric over}

\item{om_filter}{vector of OM names to calculate metric over}

\item{interval_widths}{confidence intevrals to compute}

\item{extra_filter}{an additional set of filters to apply before computing 
medians and confidence intervals}

\item{relative}{a management procedure to compute metric relative to}

\item{summarise_by}{vector of columns to summarise metric by}

\item{summary_out}{whether to output data summarised by `ggdist` or full data}
}
\description{
Compute average number of years required for SSB to recover above
0.4*SSB0 (median and CIs) following a known SSB crash
for each combination of operating models and management procedures.
}
\details{
"Recovery" begins exactly 20 years after the start of the simulation
to correspond with the "Immediate Crash Recruitment" scenario.
}
