% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{calculate_joint_selret}
\alias{calculate_joint_selret}
\title{Calculate Joint Selectivity and Retention Across Multiple Fleets}
\usage{
calculate_joint_selret(sel, ret, prop_fs = c(0.5, 0.5))
}
\arguments{
\item{sel}{selectiviity-at-age ([1, nages, nesexes, nregions, nfleets])}

\item{ret}{retention-at-age ([1, nages, nsexes, nregions, nfleets])}

\item{prop_fs}{fleet weights}
}
\description{
Computes the average selectivity-at-age and retention-at-age acting
on a population when multiple fleets are present. Selectivity and
retention are weighted based on user supplied weights.
}
