% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{calculate_realized_samplesize}
\alias{calculate_realized_samplesize}
\title{Realized Sample Size}
\usage{
calculate_realized_samplesize(ac_obs, naa, selex)
}
\arguments{
\item{ac_obs}{age composition observations}

\item{naa}{population age structure}

\item{selex}{selectivity-at-age}
}
\value{
vector of realized sample sizes for each year
}
\description{
Wrapper function for calculating realized sample size from Hulson
and williams, 2024.
}
