% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reference_points.R
\name{calculate_ref_points}
\alias{calculate_ref_points}
\title{Calculate NPFMC groundfish reference points}
\usage{
calculate_ref_points(nages, mort, mat, waa, sel, ret, avg_rec)
}
\arguments{
\item{nages}{number of ages in age structure}

\item{mort}{instantaneous natural mortality rate}

\item{mat}{maturity-at-age vector}

\item{waa}{weight-at-age vector}

\item{sel}{total selectivity-at-age vector}

\item{ret}{total retention-at-age vector}

\item{avg_rec}{average recruitment}

\item{F}{instantenous fishing mortality rate}
}
\value{
list of F40, F35, and B40
}
\description{
Calculate F_OFL (F_35%), F_ABC (F_40%), and B40
(SSB under F_ABC).
}
