% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_processing.R
\name{check_em_convergence_diagnostics}
\alias{check_em_convergence_diagnostics}
\title{Check EM Convergence Diagnostics}
\usage{
check_em_convergence_diagnostics(
  model_runs,
  extra_columns,
  hcr_filter,
  om_filter,
  n_proj_years,
  nsims
)
}
\arguments{
\item{model_runs}{list of completed MSE model runs with `diagnostics=TRUE`}

\item{extra_columns}{additional columns to associated model with HCR and OM}

\item{hcr_filter}{vector of HCR names to process (must match names in `extra_columns`)}

\item{om_filter}{vector of OM names to process (must match names in `extra_columns`)}

\item{n_proj_years}{number of projection years to compute convergence diagnostics over}

\item{nsims}{number of simulations over which to calculate convergence diagnostics}
}
\value{
dataframe of whether EM run for projection year y, and simulation s, converged
        according to the above diagnostic checks.
}
\description{
Checks a returned EM object for convergence based on the following properties:
- No infinite log-likelihood components
- Maximum gradient < 1e-3
- PD hessian
- No inifinte SEs of parameters
- No SEs > 100
- No parameter correlations > 0.999
}
