% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/age_structure_stats.R
\name{compute_average_age}
\alias{compute_average_age}
\title{Compute average age of a population}
\usage{
compute_average_age(naa, ages)
}
\arguments{
\item{naa}{numbers-at-age vector}

\item{ages}{vector of ages}
}
\description{
Calculate the average age of a population
given its age structure.
}
