% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reference_points.R
\name{compute_bx}
\alias{compute_bx}
\title{Compute average SSB under a level of F}
\usage{
compute_bx(nages, mort, mat, waa, sel, ret, F, avg_rec)
}
\arguments{
\item{nages}{number of ages in age structure}

\item{mort}{instantaneous natural mortality rate}

\item{mat}{maturity-at-age vector}

\item{waa}{weight-at-age vector}

\item{sel}{total selectivity-at-age vector}

\item{ret}{total retention-at-age vector}

\item{F}{instantenous fishing mortality rate}

\item{avg_rec}{average recruitment}
}
\description{
Compute average SSB under a given level of
fishing mortality.
}
