% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reference_points.R
\name{compute_sbpr}
\alias{compute_sbpr}
\title{Compute Spawning Biomass per Recruit (SBPR)}
\usage{
compute_sbpr(nages, mort, mat, waa, sel, ret, F)
}
\arguments{
\item{nages}{number of ages in age structure}

\item{mort}{instantaneous natural mortality rate}

\item{mat}{maturity-at-age vector}

\item{waa}{weight-at-age vector}

\item{sel}{total selectivity-at-age vector}

\item{ret}{total retention-at-age vector}

\item{F}{instantenous fishing mortality rate}
}
\description{
Compute SBPR under a given level of fishing mortality.
}
