% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/harvest_control_rules.R
\name{constant_F}
\alias{constant_F}
\title{Compute F from a constant F rule}
\usage{
constant_F(F)
}
\arguments{
\item{F}{fishing mortality rate to use}
}
\value{
fishing mortality rate
}
\description{
Compute F from a constant F rule
}
