% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{extend_3darray_last_dim}
\alias{extend_3darray_last_dim}
\title{Extend a 3D array along its final dimension}
\usage{
extend_3darray_last_dim(array_3d, n)
}
\arguments{
\item{array_3d}{array with 3 dimensions}

\item{n}{number of times to concatenate the last element in 
the 3rd dimension}
}
\description{
Copies and append values from the final dimension 
of a 3D array to the end of the final dimension.
Originally written by Craig Marsh.
}
