% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{extend_vec_last_val}
\alias{extend_vec_last_val}
\title{Extend a vector by copying its final value}
\usage{
extend_vec_last_val(vector, n)
}
\arguments{
\item{vector}{a vector to add n elements to}

\item{n}{number of times to concatenate final element}
}
\description{
Copies and appends the final value of a vector
to the end of said vector 'n' times.
}
