% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_utils.R
\name{filter_hcr_om}
\alias{filter_hcr_om}
\title{Filter processed MSE data to a subset of HCRs and OMs}
\usage{
filter_hcr_om(data, hcrs, oms)
}
\arguments{
\item{data}{dataframe of processed MSE data}

\item{hcrs}{vector of HCR names to retain in the data}

\item{oms}{vector of OM names to retain in the data}
}
\description{
Filter dataframe of processed MSE data to only include
a subset of HCRs and OMs.
}
