% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_TMB_model.R
\name{fit_TMB_model}
\alias{fit_TMB_model}
\title{Fit the SpatialSablefishAssessment TMB Model
#'
Use Template Model Builder (TMB) to fit the assessmnte model
from `SpatialSablefishAssessment`. Note that the 'ln_M_year_devs'
and 'ln_M_age_devs' parameters have been turned off.}
\usage{
fit_TMB_model(
  data,
  parameters,
  model_name = "CurrentAssessment",
  do_newton_steps = FALSE,
  fix_pars = NA,
  recompile_model = FALSE
)
}
\arguments{
\item{data}{data list to be provided to TMB (from `format_em_data(...)`)}

\item{parameters}{parameter list to be provided to TMB (from `format_em_data(...)`)}
}
\description{
Fit the SpatialSablefishAssessment TMB Model
#'
Use Template Model Builder (TMB) to fit the assessmnte model
from `SpatialSablefishAssessment`. Note that the 'ln_M_year_devs'
and 'ln_M_age_devs' parameters have been turned off.
}
