% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/catch_apportionment.R
\name{fixed_apportionment}
\alias{fixed_apportionment}
\title{Fixed Regional ABC Apportionment}
\usage{
fixed_apportionment(app, ...)
}
\arguments{
\item{app}{fixed regional apportionment (dim [1, nregions])}

\item{...}{extra posisble parameters for generalized compatability
with other apportionment functions}
}
\value{
vector of regional apportionment
}
\description{
Set a temporally fixed regional ABC apportionment.
}
