% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_em_data.R
\name{format_em_data}
\alias{format_em_data}
\title{Format OM Observation Data for TMB Assessment Model
#'
Format data and observations from an `afscOM` operating model
and update the model data and parameters required by the 
`SpatialSablefishAssessment` TMB assessment model.}
\usage{
format_em_data(
  nyears,
  dem_params,
  land_caa,
  survey_indices,
  fxfish_caa_obs,
  ll_ac_obs,
  model_options,
  added_years = 1
)
}
\arguments{
\item{nyears}{the number of years of data that will be passed to the model}

\item{dem_params}{demographic parameter matrices subsetted to 1 year}

\item{land_caa}{one year worth of landed catch-at-age data (dimensins [1, nages, nsexes, nregions, nfleets])}

\item{survey_indices}{one year worth of survey indices (LL RPN, LL RPW, and TW RPW)}

\item{fxfish_caa_obs}{one year worth of catch-at-age observation from the fixed gear fishery}

\item{ll_ac_obs}{one year worth of age composition observations frmo the longline survey}

\item{model_options}{list of model options provided to the OM}

\item{added_year}{the number of new years of data being added (should usually be 1)}
}
\description{
This function requires the existence of 'data/sablefish_em_data_2022.RDS',
and 'data/sablefish_em_par_2022.RDS' files, which come packaged with the
`SablefishMSE` codebase.
}
