% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attainment_utils.R
\docType{data}
\name{full_utilization}
\alias{full_utilization}
\title{Full TAC Utilization}
\format{
An object of class \code{matrix} (inherits from \code{array}) with 5 rows and 2 columns.
}
\usage{
full_utilization
}
\description{
Fully TAC utilization by the fleets within each region
}
\keyword{datasets}
