% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregate_observations.R
\name{generate_aggregate_comps}
\alias{generate_aggregate_comps}
\title{Generate Regionally Weighted Aggregated Age Compositions}
\usage{
generate_aggregate_comps(dem_params, naa, caa, obs_pars)
}
\arguments{
\item{dem_params}{list of demographic parameters from the OM [1, nages, nsexes, nregions, nfleets]}

\item{naa}{array of numbers at age [1, nages, nsexes, nregions]}

\item{caa}{array of catch at age [1, nages, nsexes, nregions}

\item{obs_pars}{list of observation parameters from OM}
}
\value{
array of age compositions [1, nages, nsexes, 1, nfleets+surveys]
}
\description{
Create a single-region aggregate age composition from a spatially-explicit
OM. Regional age-compositions are weighted by modifying the ISS to the
random multinomial draw based on the proportion of stock biomass or total
landings in each region. This deviates slightly from the operational method,
where regional weights come from the LL RPW index and CAA in numbers of
inidviduals.
}
