% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{generate_annual_frequency}
\alias{generate_annual_frequency}
\title{Generate frequency vector}
\usage{
generate_annual_frequency(frequency, len)
}
\arguments{
\item{frequency}{freqency with which a 1 occurs}

\item{len}{length of output vector}
}
\description{
Create a vector of 0s and 1s that that alternate with
a given frequency. This is intended to be used to
indicate when certain actions should occur within a 
loop.
}
