% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_RTMB_inputs.R
\name{generate_RTMB_inputs}
\alias{generate_RTMB_inputs}
\title{Generate Data, Parameter, and Mapping Inputs for SPoRK Estimation Model}
\usage{
generate_RTMB_inputs(
  nyears,
  dem_params,
  agg_land_caa,
  aggregated_survey_obs,
  model_options,
  r_ISS
)
}
\arguments{
\item{nyears}{number of years data observations}

\item{dem_params}{list of demographic parameters}

\item{agg_land_caa}{aggregated landed catch at age array [nyears, nages, nsexes, 1, nfleets]}

\item{aggregated_survey_obs}{list of aggregated survey observations (indices and age compositions)}

\item{model_options}{OM model options}

\item{r_ISS}{input sample sizes for age composition data [nyears, nfleets+nsurveys]}
}
\value{
list of data, parameters, and parameter mappings for SPoRK TMB model
}
\description{
Generate Data, Parameter, and Mapping Inputs for SPoRK Estimation Model
}
