% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_processing.R
\name{get_atage_groups}
\alias{get_atage_groups}
\title{Get Catch or Numbers-at-Age by Age Groups}
\usage{
get_atage_groups(
  model_runs,
  extra_columns,
  hcr_filter,
  om_filter,
  q,
  age_groups,
  group_names,
  group_abbs,
  spinup_years = 64,
  summarise = FALSE,
  make_segments = FALSE
)
}
\arguments{
\item{model_runs}{list of completed MSE simulation objects}

\item{extra_columns}{additional columns to append to output}

\item{q}{either "caa" (for catch-at-age) or "naa" (for numbers-at-age)}

\item{age_groups}{ages that define age groups (3 groups required)}

\item{group_names}{names for each age group}

\item{group_abbs}{abbreviated names for each age group}

\item{summarise}{whether to summarise across simualtions or not}

\item{make_segments}{whether to generate data.frame of segment for use in plotting}
}
\description{
Process MSE simulation results for -at-age by
specified age groups.
}
