% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_processing.R
\name{get_b40_timeseries}
\alias{get_b40_timeseries}
\title{Get Timeseries of B40 from MSE Simulations}
\usage{
get_b40_timeseries(model_runs, extra_columns, hcr_filter, om_filter)
}
\arguments{
\item{model_runs}{list of completed MSE simulations}

\item{extra_columns}{additional columns that should be
appended to the resultant data frame}

\item{hcr_filter}{vector of HCR names to process (must match names in `extra_columns`)}

\item{om_filter}{vector of OM names to process (must match names in `extra_columns`)}
}
\value{
timeseries of B40 reference point summarised across simulations
}
\description{
Derive timeseries of B40 from completed MSE
simulations.
}
