% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_processing.R
\name{get_fishing_mortalities}
\alias{get_fishing_mortalities}
\title{Get Annual Fishing Mortality}
\usage{
get_fishing_mortalities(model_runs, extra_columns, hcr_filter, om_filter)
}
\arguments{
\item{model_runs}{list of completed MSE simulation objects}

\item{extra_columns}{additional columns to append to output}

\item{hcr_filter}{vector of HCR names to process (must match names in `extra_columns`)}

\item{om_filter}{vector of OM names to process (must match names in `extra_columns`)}
}
\description{
Process MSE simulations for fishing mortality by fleet.
Total fishing mortality across fleets is also computed as the
sum of fleet-specific fishing mortality.
}
