% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_processing.R
\name{get_landed_catch}
\alias{get_landed_catch}
\title{Get Landed Catches}
\usage{
get_landed_catch(model_runs, extra_columns, hcr_filter, om_filter)
}
\arguments{
\item{model_runs}{list of completed MSE simulation objects}

\item{extra_columns}{additional columns to append to output}

\item{hcr_filter}{vector of HCR names to process (must match names in `extra_columns`)}

\item{om_filter}{vector of OM names to process (must match names in `extra_columns`)}
}
\description{
Process MSE simulations for landed catches by fleet.
Total landed catch across fleets is also computed.
}
