% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_processing.R
\name{get_management_quantities}
\alias{get_management_quantities}
\title{Get ABC, TAC, and Expected Landings}
\usage{
get_management_quantities(
  model_runs,
  extra_columns,
  hcr_filter,
  om_filter,
  spinup_years = 54
)
}
\arguments{
\item{model_runs}{list of completed MSE simulation objects}

\item{extra_columns}{additional columns to append to output}

\item{hcr_filter}{vector of HCR names to process (must match names in `extra_columns`)}

\item{om_filter}{vector of OM names to process (must match names in `extra_columns`)}

\item{spinup_years}{year at which to begin calculating quantities}
}
\description{
Process MSE simulations for ABC, TAC, and expected 
landings quantities. Historical ABCs, TACs, and landings (e.g., from the
conditioning period) are not available.
}
