% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_processing.R
\name{get_phaseplane_data}
\alias{get_phaseplane_data}
\title{Get F and SSB data for Phaseplane Plot}
\usage{
get_phaseplane_data(
  model_runs,
  extra_columns,
  dem_params,
  hcr_filter,
  om_filter
)
}
\arguments{
\item{model_runs}{list of completed MSE simulations}

\item{extra_columns}{additional columns that should be
appended to the resultant data frame}

\item{dem_params}{list of demographic parameters values}

\item{hcr_filter}{vector of HCR names to process (must match names in `extra_columns`)}

\item{om_filter}{vector of OM names to process (must match names in `extra_columns`)}
}
\description{
Process MSE simulation data for total F
and SSB by year across simulations, OMs,
and management procedures.
}
