% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_processing.R
\name{get_reference_points}
\alias{get_reference_points}
\title{Get Reference Points from MSE Simulations}
\usage{
get_reference_points(
  model_runs,
  extra_columns,
  hcr_filter,
  om_filter,
  seed_list
)
}
\arguments{
\item{model_runs}{list of completed MSE simulations}

\item{extra_columns}{additional columns that should be appended to the resultant data frame}

\item{hcr_filter}{vector of HCR names to process (must match names in `extra_columns`)}

\item{om_filter}{vector of OM names to process (must match names in `extra_columns`)}

\item{seed_list}{simulation seeds used in `model_runs`}
}
\description{
Derive fishing mortality and biomass reference points
from completed MSE simulations.
}
