% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_processing.R
\name{get_rp_timeseries}
\alias{get_rp_timeseries}
\title{Get Timeseries of Refernce Point from MSE Simulations}
\usage{
get_rp_timeseries(
  model_runs,
  extra_columns,
  hcr_filter,
  om_filter,
  ref_pt,
  spr_target = "hcr"
)
}
\arguments{
\item{model_runs}{list of completed MSE simulations}

\item{extra_columns}{additional columns that should be
appended to the resultant data frame}

\item{hcr_filter}{vector of HCR names to process (must match names in `extra_columns`)}

\item{om_filter}{vector of OM names to process (must match names in `extra_columns`)}

\item{ref_pt}{the reference point to return (one of "Fmax", "Fref", "Bref", or "B0")}

\item{spr_target}{target SPR reference level to compute timeseries of (if "hcr", will calculate
uniquely for each HCR based on the the SPR target reference point defined in the HCR object)}
}
\value{
timeseries of reference point value
}
\description{
Derive timeseries of a given reference point from completed MSE
simulations.
}
