% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_saved_model_runs}
\alias{get_saved_model_runs}
\title{Load Saved MSE Model Runs from Disk}
\usage{
get_saved_model_runs(om_order = NULL, hcr_order = NULL)
}
\arguments{
\item{om_order}{vector of correct order of OMs (used to set OM factor level)}

\item{hcr_order}{vector of correct order of HCRs (used to set HCR factor level)}
}
\description{
Read all saved RDS files present in data/active and coerce into proper
model_runs list object. Also setup correctly specified extra_columns
object for use with bind_mse_outputs.
}
