% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_processing.R
\name{get_ssb_biomass}
\alias{get_ssb_biomass}
\title{Get Spawning Biomass and Total Biomass}
\usage{
get_ssb_biomass(model_runs, extra_columns, dem_params, hcr_filter, om_filter)
}
\arguments{
\item{model_runs}{list of completed MSE simulation objects}

\item{extra_columns}{additional columns to append to output}

\item{dem_params}{demographic parameter matrix}

\item{hcr_filter}{vector of HCR names to process (must match names in `extra_columns`)}

\item{om_filter}{vector of OM names to process (must match names in `extra_columns`)}
}
\description{
Process MSE simulations for spawning biomass,
and total stock biomass.
}
