% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{inf_max}
\alias{inf_max}
\title{Get maximum value without considering infinite values}
\usage{
inf_max(d)
}
\arguments{
\item{d}{vector of values to find maximum of}
}
\description{
Wrapper around max that ignores infinite values
}
