% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recruitment_utils.R
\name{multivariate_recruit_apportionment}
\alias{multivariate_recruit_apportionment}
\title{Generate Random Spatial Recruitment Apportionment matrix}
\usage{
multivariate_recruit_apportionment(seed)
}
\arguments{
\item{seed}{random seed for reproducability}

\item{nyears}{total number of years to simulate}
}
\description{
Use a mutilvariate normal distribution to generate a random vector/matrix
of spatial recruitment apportionment, while preserving historical correlations
and covariances through a variance-covariance matrix.
}
