% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recruitment_utils.R
\name{multivariate_recruits}
\alias{multivariate_recruits}
\title{Generate Random Spatial Recruitment matrix}
\usage{
multivariate_recruits(nyears, seed)
}
\arguments{
\item{nyears}{total number of years to simulate}

\item{seed}{random seed for reproducability}
}
\description{
Use a mutilvariate normal distribution to generate a random vector/matrix
of spatial recruitment events, while preserving historical correlations
and covariances through a variance-covariance matrix.
}
