% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/harvest_control_rules.R
\name{npfmc_tier3_F}
\alias{npfmc_tier3_F}
\title{Compute F from Tier 3 HCR}
\usage{
npfmc_tier3_F(ssb, B_ref, F_ref)
}
\arguments{
\item{ssb}{Value for SSB}

\item{B_ref}{Value for biomass based reference point (e.g., B40)}

\item{F_ref}{Value for fishing mortality based reference point (e.g., F40)}
}
\value{
fishing mortality rate
}
\description{
Compute F from Tier 3 HCR
}
