% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_alaska_map}
\alias{plot_alaska_map}
\title{Plot Map of Alaska with NMFS Stat Areas Colored}
\usage{
plot_alaska_map(regions = c("BS", "AI", "WGOA", "CGOA", "EGOA"))
}
\description{
Original code from Matt Cheng
}
