% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_em_diagnostics}
\alias{plot_em_diagnostics}
\title{Plot EM Diagnostics}
\usage{
plot_em_diagnostics(
  model_runs,
  extra_columns,
  hcr_filter,
  om_filter,
  spinup_years,
  n_proj_years,
  simulation_year = 1,
  simulation_number = 1
)
}
\arguments{
\item{model_runs}{list of completed MSE simulation objects}

\item{extra_columns}{additional columns to append to output}

\item{hcr_filter}{vector of HCR names to process (should only be length 1)}

\item{om_filter}{vector of OM names to process (should only be length 1)}

\item{spinup_years}{Number of years of model outputs prior to start of EM (mse_options$n_spinup_years)}

\item{n_proj_years}{Number of years of EM model outputs (mse_options$n_proj_years)}

\item{simulation_year}{Year of the simulation to plot diagnostics for}

\item{simulation_number}{Simulation number to plot diagnostics for}
}
\value{
ggplot figure showing model diagnostics plots
}
\description{
Take model runs and plot set of model dignostics, including:
1) Model fits to catch and indices
2) Model fits to age composition data
3) Model fits to derived quantities (SSB, harvest rate, recruitment)
}
