% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/performance_metrics.R
\name{prop_low_biomass}
\alias{prop_low_biomass}
\title{Compute proportion of years SSB is below a threshold level across projection period}
\usage{
prop_low_biomass(
  model_runs,
  extra_columns,
  dem_params,
  hcr_filter,
  om_filter,
  interval_widths = c(0.5, 0.8),
  time_horizon = c(55, NA),
  extra_filter = NULL,
  relative = NULL,
  summarise_by = c("om", "hcr"),
  summary_out = TRUE
)
}
\arguments{
\item{model_runs}{list of completed MSE simulations runs}

\item{extra_columns}{data.frame specifying names for OM and HCR to attach
to each model_run (see `bind_mse_outputs` for more details)}

\item{hcr_filter}{vector of HCR names to calculate metric over}

\item{om_filter}{vector of OM names to calculate metric over}

\item{interval_widths}{confidence intevrals to compute}

\item{time_horizon}{periof of years over which to calculate metric}

\item{extra_filter}{an additional set of filters to apply before computing 
medians and confidence intervals}

\item{relative}{a management procedure to compute metric relative to}

\item{summarise_by}{vector of columns to summarise metric by}

\item{summary_out}{whether to output data summarised by `ggdist` or full data}
}
\description{
Compute average proportion of years where SSB is below 0.4*SSB0
(median and CIs) across all years and simulation seeds, 
for each combination of operating models and management procedures.
}
