% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/age_structure_stats.R
\name{prop_mature}
\alias{prop_mature}
\title{Compute proportion of a population that is mature}
\usage{
prop_mature(naa, mat)
}
\arguments{
\item{naa}{numbers-at-age vector}

\item{mat}{maturity-at-age vector}
}
\description{
Calculate the proportion of a population that
has reached maturity given its age structure
and maturity schedule.
}
