% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recruitment_utils.R
\name{recruits_crash}
\alias{recruits_crash}
\title{Recruitment function with a specified crash period}
\usage{
recruits_crash(
  crash_start_year,
  crash_length,
  crash_value,
  hist_recruits,
  nyears,
  seed
)
}
\arguments{
\item{crash_length}{number of years to have "crashed" recruitment}

\item{crash_value}{recruitment during the "crash" period}

\item{hist_recruits}{historical timeseries of recruitment (or deviates)}

\item{nyears}{total number of years to resample}

\item{seed}{random seed for reproducability}

\item{crash_start_years}{projection year to start the recruitment crash}
}
\description{
Generate future recruitment timeseries via historical reasampling, 
(see `resample_recruits`) but set a specific time period of the
projected recruitment to a different values. Can be used to specify
a "crash" in recruitment, or a "spike" in recruitment depending on
the `crash_value`.
}
