% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_utils.R
\name{reformat_ggdist_long}
\alias{reformat_ggdist_long}
\title{Change a `ggdist` Tibble to Long Format}
\usage{
reformat_ggdist_long(data, n = 1)
}
\arguments{
\item{data}{a tibble output by a `ggdist` summary function
(e.g. `ggdist:median_qi(...)`)}

\item{n}{the number of grouping variables used (default: 1)}
}
\value{
a long format tibble with columns median, lower, and upper
}
\description{
Pivot a summary tibble from `ggdist` with multiple summary variables
to long format. Summary variable names will be be pivoted to a 
'name' column, and corresponding 'median', 'lower', 'upper' columns
will contain computed quantile values.
}
\details{
Intended for use when plotting distributions of multiple variables
simultaneously.
}
