% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recruitment_utils.R
\name{regime_recruits}
\alias{regime_recruits}
\title{Regime-like recruitment}
\usage{
regime_recruits(mus, cvs, nyears, regime_length, starting_regime, seed)
}
\arguments{
\item{mus}{vector of mean recruitment in each regime}

\item{cvs}{vector of CV of recruitment in each regime (CVs are converted internally to log SD)}

\item{nyears}{total number of years to resample}

\item{regime_length}{vector inidcating how long each regime last (limited to 2 regimes)}

\item{starting_regime}{regime to start with (0 = first regime, 1 = second regime)}

\item{seed}{random seed for reproducability}
}
\description{
Generate future recruitment timeseries based on mean and CV of distinct
regimes, via a lognormal distribution.
}
