% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recruitment_utils.R
\name{resample_recruit_apportionment}
\alias{resample_recruit_apportionment}
\title{Resample from Historical Spatial Recruit Apportionment}
\usage{
resample_recruit_apportionment(hist_recruits, seed)
}
\arguments{
\item{hist_recruits}{historical timeseries of recruitment (or deviates)}

\item{seed}{random seed for reproducability}
}
\description{
Resample, with replacement, from the timeseries of historical
regional recruit apportionment.
}
