% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recruitment_utils.R
\name{resample_recruits}
\alias{resample_recruits}
\title{Resample recruitment from historical timeseries
#'
Resample, with replacement, the historical recruitment timeseries.}
\usage{
resample_recruits(hist_recruits, nyears, seed)
}
\arguments{
\item{hist_recruits}{historical timeseries of recruitment (or deviates)}

\item{nyears}{total number of years to resample}

\item{seed}{random seed for reproducability}
}
\description{
Resample recruitment from historical timeseries
#'
Resample, with replacement, the historical recruitment timeseries.
}
