% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recruitment_utils.R
\name{resample_recruits_spatial}
\alias{resample_recruits_spatial}
\title{Resample recruitment from historical spatial recruits timeseries}
\usage{
resample_recruits_spatial(hist_recruits, nyears, seed)
}
\arguments{
\item{hist_recruits}{historical timeseries of recruitment (or deviates)}

\item{nyears}{total number of years to resample}

\item{seed}{random seed for reproducability}
}
\description{
Resample, with replacement, from the historical recruitment timeseries.
}
