% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recruitment_utils.R
\name{resample_regime_recruits}
\alias{resample_regime_recruits}
\title{Regime-like recruitment
#'
Generate future recruitment timeseries based on historcial regime-like 
recruitment dynamics. Historical recruitments identified as having belonged
to a specific regime are randomly resampled, with replacement.}
\usage{
resample_regime_recruits(
  regime1_recruits,
  regime2_recruits,
  nyears,
  regime_length,
  starting_regime,
  seed
)
}
\arguments{
\item{regime1_recruits}{historical recruitments from first regime}

\item{regime2_recruits}{historical recruitments from second regime}

\item{nyears}{total number of years to resample}

\item{regime_length}{vector inidcating how long each regime last (limited to 2 regimes)}

\item{starting_regime}{regime to start with (0 = first regime, 1 = second regime)}

\item{seed}{random seed for reproducability}
}
\description{
Regime-like recruitment
#'
Generate future recruitment timeseries based on historcial regime-like 
recruitment dynamics. Historical recruitments identified as having belonged
to a specific regime are randomly resampled, with replacement.
}
