% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_utils.R
\name{round_to_zero}
\alias{round_to_zero}
\title{Round dataframe column to Zero}
\usage{
round_to_zero(data, col_name, zero_threshold = 0.01)
}
\arguments{
\item{data}{dataframe of processed MSE data}

\item{col_name}{name of data column to round}

\item{zero_threshold}{threshold to round down to zero}
}
\description{
Rounds data in a dataframe column to 0 when below a 
certain threshold. Also rounds non finite values to 
0 automatically.
}
