% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/catch_apportionment.R
\name{rpw_moving_average}
\alias{rpw_moving_average}
\title{Regional ABC Apportionment by RPW Moving Average}
\usage{
rpw_moving_average(survey_obs, y, window_size = 5)
}
\arguments{
\item{survey_obs}{list of survey observations with an RPW list element}

\item{y}{the terminal year of observations to calculate the moving average over}

\item{window_size}{size of the moving average window to use}
}
\value{
proportion of survey RPW in each spatial region in year y
}
\description{
The operational method for regional ABC apportionment for sablefish
in Alaska is a simple 5-yr moving average across the Alaska LL survey
RPW index.
}
