% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{rss}
\alias{rss}
\title{Realized Sample Size}
\usage{
rss(obs, exp)
}
\arguments{
\item{obs}{age composition observations}

\item{exp}{expected age composition}
}
\value{
realized sample size sum(obs*(1-obs))/sum((exp-obs)^2)
}
\description{
Calculate realized sample size from Hulson and Williams, 2024
}
