% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_em.r
\name{run_EM}
\alias{run_EM}
\title{Run Estimation Method}
\usage{
run_EM(
  s,
  sable_om,
  mse_tier3,
  nyears,
  om_name,
  model_name = "CurrentAssessment",
  fix_pars = NA
)
}
\arguments{
\item{s}{simulation seed}

\item{sable_om}{om object}

\item{mse_tier3}{mse object}

\item{om}{string identifed for OM}
}
\description{
Given an OM object and a succesfull MSE run, apply
the EM to the observation data generated by the MSE.
}
