% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_mse.R
\name{run_mse}
\alias{run_mse}
\title{Run Management Strategy Evaluation}
\usage{
run_mse(om, mp, mse_options, seed = 1120)
}
\arguments{
\item{om}{an operating model list object (should minimally contain a
list object with demographic parameters, an array of initial NAAs, a
list object specifying recruitment, and a list of model_options).}

\item{mp}{a list object specifying the management procedure to be applied during
the projection period}

\item{mse_options}{a list object specified MSE options}

\item{seed}{random seed}
}
\description{
Run an MSE simulation loop.
}
