% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_mse_parallel.R
\name{run_mse_parallel}
\alias{run_mse_parallel}
\title{Run Multiple MSE Simulations in Parallel}
\usage{
run_mse_parallel(nsims, seeds, om, hcr, mse_options, diagnostics = FALSE, ...)
}
\arguments{
\item{nsims}{number of simulations to run}

\item{seeds}{simulation seeds to use}

\item{om}{operating model object}

\item{hcr}{management procedure object}

\item{mse_options}{MSE options object}

\item{diagnostics}{whether to return EM model objects for diagnostic testing}
}
\value{
list of MSE outputs (NAA, CAA, FAA, etc.)
}
\description{
Run Multiple MSE Simulations in Parallel
}
