% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sable_om}
\alias{sable_om}
\title{Default Operating Model for Alaska Sablfish}
\format{
list containing OM components
}
\usage{
sable_om
}
\description{
Dataset containing demographic matrices, observation
process parameters, and initial numbers-at-age for 
the alaska sablefish population. This data is used as
the base for all operating models used in this 
package.
}
\keyword{datasets}
