% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_utils.R
\name{scale_and_rank}
\alias{scale_and_rank}
\title{Scale and Rank dataframe columns within groups}
\usage{
scale_and_rank(data, col_name)
}
\arguments{
\item{data}{dataframe of processed MSE data}

\item{col_name}{name of data column to scale and rank}
}
\description{
Computes 0-1 scaled version of dataframe column, within
groups, and ranks (1-N) entries within group based on
scaled values (where 1 is highest in-group scaled value, and
N is the lowest in-group scaled value).
}
\details{
Intended for use on performance metric data only. Note that
"Catch AAV, "Proportion of Years with Low SSB", and "Recovery
Time" performance metrics use inverse ranking system (smaller
scaled values are better), as the intention is to minimize
these metrics.
}
