% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/age_structure_stats.R
\name{shannon_diversity}
\alias{shannon_diversity}
\title{Compute Shannon Diversity Index (H')}
\usage{
shannon_diversity(naa)
}
\arguments{
\item{naa}{numbers-at-age vector}
}
\description{
Calculate the Shannon diversity index (H')
(Shannon and Weaver, 1949) given a numbers
at age vector.
}
\details{
H' = sum(naa*log(naa))
}
