% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reference_points.R
\name{spr_x}
\alias{spr_x}
\title{Find F that yields a given SPR%}
\usage{
spr_x(nages, mort, mat, waa, sel, ret, target_x = 0.35)
}
\arguments{
\item{nages}{number of ages in age structure}

\item{mort}{instantaneous natural mortality rate}

\item{mat}{maturity-at-age vector}

\item{waa}{weight-at-age vector}

\item{sel}{total selectivity-at-age vector}

\item{ret}{total retention-at-age vector}

\item{target_x}{desired SPR proportion}
}
\description{
Use bisection algorithm to identify the level of
fishing mortality required to yield an SPR of x%.
}
