% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attainment_utils.R
\name{stairstep_attainment}
\alias{stairstep_attainment}
\title{Stairstep Attainment Function}
\usage{
stairstep_attainment(v, breakpoints, levels, phase_ins)
}
\arguments{
\item{v}{a total allowable catch (TAC) level}

\item{breakpoints}{TAC levels that define where a transition
in attainment level occurs}

\item{levels}{attainment levels associated with each set of breakpoints
(0 - breakpoint[1], breakpoinut[1]-breakpoint[2], breakpoint[2] - Inf)}

\item{phase_ins}{optional phase_in width. When specified, a phase ins
begin talking effect at phase_ins/2 before each breakpoints value and
end at phase_ins/2 after each breakpoint value.}
}
\description{
Defines attainment level based on a three stepped function,
with optional phase ins periods where a linear combination of
stairsteps is applied. Attainment is assumed to be a function
of the established TAC.
}
