% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/harvest_control_rules.R
\name{threshold_f}
\alias{threshold_f}
\title{Generic threshold F harvest control rule
#'
A generic 4 parameters threshold F HCR.}
\usage{
threshold_f(x, f_min, f_max, lrp, urp)
}
\arguments{
\item{x}{input to HCR (e.g. SSB or, SSB/B40)}

\item{f_min}{F to use when x < lrp}

\item{f_max}{F to use when x > urp}

\item{lrp}{value of x below which f_min applies}

\item{urp}{value of x above which f_max applies}
}
\value{
fishing mortality rate
}
\description{
Generic threshold F harvest control rule
#'
A generic 4 parameters threshold F HCR.
}
