% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reference_points.R
\name{calculate_ref_points}
\alias{calculate_ref_points}
\title{Calculate SPR Based Reference Points}
\usage{
calculate_ref_points(nages, mort, mat, waa, sel, ret, avg_rec, spr_target)
}
\arguments{
\item{nages}{number of ages in age structure}

\item{mort}{instantaneous natural mortality rate}

\item{mat}{maturity-at-age vector}

\item{waa}{weight-at-age vector}

\item{sel}{total selectivity-at-age vector}

\item{ret}{total retention-at-age vector}

\item{avg_rec}{average recruitment}

\item{spr_target}{target SPR level. If length(spr_target) == 1, same
target SPR level assumed for the F and B reference points. If 
length(spr_target) == 2, element 1 corresponds to the spr target for 
F, while element 2 corresponds to the spr target for B.}
}
\value{
list of F reference point and biomass reference point
}
\description{
Generalized function for computing and F and biomass based reference point
for a given SPR level
}
