% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/age_structure_stats.R
\name{prop_fully_mature}
\alias{prop_fully_mature}
\title{Compute proportion of population that is fully mature
#'
Calculate the proportion of a population
in age classes that are "fully mature" (e.g.
mat > 0.995).}
\usage{
prop_fully_mature(naa, mat)
}
\arguments{
\item{naa}{numbers-at-age vector}

\item{mat}{maturity-at-age vector}
}
\description{
Compute proportion of population that is fully mature
#'
Calculate the proportion of a population
in age classes that are "fully mature" (e.g.
mat > 0.995).
}
