% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_utils.R
\name{reformat_ggdist_long}
\alias{reformat_ggdist_long}
\title{Change a `ggdist` Tibble to Long Format 
#'
Description}
\usage{
reformat_ggdist_long(data, n = 1)
}
\arguments{
\item{data}{a tibble output by a `ggdist` summary function
(e.g. `ggdist:median_qi(...)`)}

\item{n}{the number of grouping variables used (default: 1)}
}
\value{
a long format tibble with columns median, lower, and upper
}
\description{
Change a `ggdist` Tibble to Long Format 
#'
Description
}
