% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_mse.R
\name{run_mse}
\alias{run_mse}
\title{Run Management Strategy Evaluation
#'
Run an MSE simulation loop.}
\usage{
run_mse(om, hcr, ..., nyears_input = NA, spinup_years = 64, seed = 1120)
}
\arguments{
\item{om}{path to a saved OM .RDS file (should contain
demagraphic parameters list and model options list)}

\item{hcr}{a function to compute the allowable F in the next year}

\item{...}{parameters to pass to the `hcr` function}

\item{nyears_input}{number of years to simulate forward (will override the dimensions
of the demagraphic parameters matrices defined in the OM)}

\item{spinup_years}{number of years before estimation process should begin}

\item{seed}{random seed}
}
\description{
Run Management Strategy Evaluation
#'
Run an MSE simulation loop.
}
