% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_mse_multiple.R
\name{run_mse_multiple}
\alias{run_mse_multiple}
\title{Run MSE Simulations Across OMs and HCRs}
\usage{
run_mse_multiple(
  om_list,
  hcr_list,
  seed_list,
  mse_options_list,
  nyears,
  diagnostics = FALSE,
  save = FALSE
)
}
\arguments{
\item{om_list}{named list of operating model list objects}

\item{hcr_list}{named list of harvest control rule list objects}

\item{seed_list}{vector of random seeds}

\item{mse_options}{list of mse_options objects to apply to each
OM. Must be either length 1, or length == length(om_list).}

\item{...}{additional parameters to pass to the `run_mse` call}
}
\value{
list of MSE simualtion results
}
\description{
Wrapper function around `run_mse` and `run_mse_parallel` that
allows for simple running of MSE simulations across multiple 
OMs and HCRs (and combinations of the two). The wrapper relies
on using `expand.grid` to create a factorial combination of OMs
and HCRs and runs MSE simulations across all such combinations.
}
