% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_mse_multiple.R
\name{run_mse_multiple}
\alias{run_mse_multiple}
\title{Run Multiple MSE Simulations
#'
Wrapper function around `run_mse` that handle running and
compiling the outputs of multiple MSE closed-loop simulations,
across different random seeds.}
\usage{
run_mse_multiple(nsims, seeds, nyears, ...)
}
\arguments{
\item{nsims}{the number of unique simulations to run}

\item{seeds}{vector of random seeds to use (length(seeds) == nsims)}

\item{nyears}{number of years in each simulation}

\item{...}{additional parameter sto pass to the `run_mse` call}
}
\value{
list of arrays of containing OM results from each simulation
}
\description{
Run Multiple MSE Simulations
#'
Wrapper function around `run_mse` that handle running and
compiling the outputs of multiple MSE closed-loop simulations,
across different random seeds.
}
